rename package with derivation
