+++ android/src/main/java/io/parity/signer/screens/createderivation/help/DerivationMethodsHelpBottomSheet.kt
@@ -1,33 +1,18 @@
-package io.parity.signer.screens.keyderivation.help
+package io.parity.signer.screens.createderivation.help
 
 import android.content.res.Configuration
 import androidx.compose.foundation.*
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Icon
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Info
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalInspectionMode
-import androidx.compose.ui.res.dimensionResource
-import androidx.compose.ui.res.pluralStringResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import io.parity.signer.R
 import io.parity.signer.components.base.BottomSheetHeader
-import io.parity.signer.components.base.SecondaryButtonWide
-import io.parity.signer.components.base.SignerDivider
-import io.parity.signer.components.qrcode.AnimatedQrKeysInfo
-import io.parity.signer.components.qrcode.EmptyAnimatedQrKeysProvider
 import io.parity.signer.models.Callback
-import io.parity.signer.models.KeySetModel
-import io.parity.signer.ui.helpers.PreviewData
 import io.parity.signer.ui.theme.*
 
 @Composable
